CHANGE: Create and update implemenation and tests. CHANGE: use BIGINT instead of int for SQL model.

+++ test_data.sql
@@ -8,10 +8,10 @@ DROP TABLE IF EXISTS `Rooms`;
 DROP TABLE IF EXISTS `Displays`;
 
 CREATE TABLE `Displays` (
-	`id` int NOT NULL AUTO_INCREMENT,
+	`id` BIGINT NOT NULL AUTO_INCREMENT,
 	`name` varchar(255) NOT NULL,
 	`utc_updated` datetime DEFAULT CURRENT_TIMESTAMP() ON UPDATE CURRENT_TIMESTAMP(),
-	`room_id` int NOT NULL,
+	`room_id` BIGINT NOT NULL,
 	PRIMARY KEY (`id`),
 	UNIQUE KEY `id_UNIQUE` (`id`)
 ) ENGINE InnoDB,
@@ -20,7 +20,7 @@ CREATE TABLE `Displays` (
 
 
 CREATE TABLE `Rooms` (
-	`id` int NOT NULL AUTO_INCREMENT,
+	`id` BIGINT NOT NULL AUTO_INCREMENT,
 	`name` varchar(255) NOT NULL,
 	`label` varchar(255),
 	`utc_updated` datetime DEFAULT CURRENT_TIMESTAMP() ON UPDATE CURRENT_TIMESTAMP(),
